move LocalBuildsRoot into separate module
